refactor: rename FullscreenLoadingIndicator
